+++ sync-for-payroll/src/codatsyncpayroll/models/operations/get_account.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import account as shared_account
-from ..shared import errormessage as shared_errormessage
+from ...models.shared import account as shared_account
 from typing import Optional
 
 
@@ -26,8 +25,6 @@ class GetAccountResponse:
     r"""HTTP response status code for this operation"""
     account: Optional[shared_account.Account] = dataclasses.field(default=None)
     r"""Success"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""