+++ sync-for-payroll/src/codatsyncpayroll/models/operations/get_journal_entry.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import journalentry as shared_journalentry
+from ...models.shared import journalentry as shared_journalentry
 from typing import Optional
 
 
@@ -24,8 +23,6 @@ class GetJournalEntryResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
     journal_entry: Optional[shared_journalentry.JournalEntry] = dataclasses.field(default=None)
     r"""Success"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)