+++ sync-for-payroll/src/codatsyncpayroll/models/operations/get_pull_operation.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import pulloperation as shared_pulloperation
+from ...models.shared import pulloperation as shared_pulloperation
 from typing import Optional
 
 
@@ -24,8 +23,6 @@ class GetPullOperationResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
     pull_operation: Optional[shared_pulloperation.PullOperation] = dataclasses.field(default=None)
     r"""OK"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)