+++ sync-for-payroll/src/codatsyncpayroll/models/operations/list_connections.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import connections as shared_connections
-from ..shared import errormessage as shared_errormessage
+from ...models.shared import connections as shared_connections
 from typing import Optional
 
 
@@ -32,8 +31,6 @@ class ListConnectionsResponse:
     r"""HTTP response status code for this operation"""
     connections: Optional[shared_connections.Connections] = dataclasses.field(default=None)
     r"""OK"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your `query` parameter was not correctly formed"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""