+++ sync-for-payroll/src/codatsyncpayroll/sdk.py
@@ -12,7 +12,7 @@ from .sdkconfiguration import SDKConfiguration
 from .tracking_categories import TrackingCategories
 from codatsyncpayroll import utils
 from codatsyncpayroll.models import shared
-from typing import Dict
+from typing import Callable, Dict, Union
 
 class CodatSyncPayroll:
     r"""Sync for Payroll: The API for Sync for Payroll.
@@ -36,27 +36,27 @@ class CodatSyncPayroll:
     | Company info         | View company profile from the source platform.                                                             |
     | Manage data          | Control how data is retrieved from an integration.                                                         |
     """
-    accounts: Accounts
-    r"""Accounts"""
     companies: Companies
     r"""Create and manage your Codat companies."""
-    company_info: CompanyInfo
-    r"""View company information fetched from the source platform."""
     connections: Connections
     r"""Manage your companies' data connections."""
+    accounts: Accounts
+    r"""Accounts"""
     journal_entries: JournalEntries
     r"""Journal entries"""
     journals: Journals
     r"""Journals"""
     manage_data: ManageData
     r"""Asynchronously retrieve data from an integration to refresh data in Codat."""
+    company_info: CompanyInfo
+    r"""View company information fetched from the source platform."""
     tracking_categories: TrackingCategories
     r"""Tracking categories"""
 
     sdk_configuration: SDKConfiguration
 
     def __init__(self,
-                 security: shared.Security = None,
+                 security: Union[shared.Security,Callable[[], shared.Security]] = None,
                  server_idx: int = None,
                  server_url: str = None,
                  url_params: Dict[str, str] = None,
@@ -66,7 +66,7 @@ class CodatSyncPayroll:
         """Instantiates the SDK configuring it with the provided parameters.
         
         :param security: The security details required for authentication
-        :type security: shared.Security
+        :type security: Union[shared.Security,Callable[[], shared.Security]]
         :param server_idx: The index of the server to use for all operations
         :type server_idx: int
         :param server_url: The server URL to use for all operations
@@ -81,25 +81,21 @@ class CodatSyncPayroll:
         if client is None:
             client = requests_http.Session()
         
-        
-        security_client = utils.configure_security_client(client, security)
-        
-        
         if server_url is not None:
             if url_params is not None:
                 server_url = utils.template_url(server_url, url_params)
 
-        self.sdk_configuration = SDKConfiguration(client, security_client, server_url, server_idx, retry_config=retry_config)
+        self.sdk_configuration = SDKConfiguration(client, security, server_url, server_idx, retry_config=retry_config)
        
         self._init_sdks()
     
     def _init_sdks(self):
-        self.accounts = Accounts(self.sdk_configuration)
         self.companies = Companies(self.sdk_configuration)
-        self.company_info = CompanyInfo(self.sdk_configuration)
         self.connections = Connections(self.sdk_configuration)
+        self.accounts = Accounts(self.sdk_configuration)
         self.journal_entries = JournalEntries(self.sdk_configuration)
         self.journals = Journals(self.sdk_configuration)
         self.manage_data = ManageData(self.sdk_configuration)
+        self.company_info = CompanyInfo(self.sdk_configuration)
         self.tracking_categories = TrackingCategories(self.sdk_configuration)