+++ sync-for-payroll/RELEASES.md
@@ -49,3 +49,13 @@ Based on:
 - [python v2.1.0] sync-for-payroll
 ### Releases
 - [PyPI v2.1.0] https://pypi.org/project/codat-sync-for-payroll/2.1.0 - sync-for-payroll
+
+## 2023-11-13 01:24:46
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payroll.yaml
+- Speakeasy CLI 1.119.0 (2.188.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.0.0] sync-for-payroll
+### Releases
+- [PyPI v3.0.0] https://pypi.org/project/codat-sync-for-payroll/3.0.0 - sync-for-payroll